add helper checks module
